+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.0: 2026-03-19
 
+* Fix month status projection to separate bills from discretionary spending
 * Change bill due text to "Erääntyy X. päivä" in Finnish
 * Fix greeting value colors to only apply to euro amounts
 * Align sidebar collapse button to left